+++ packages/in_app_purchase/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.1.1
 
+* Fix new lint warnings.
+* Update minimum Flutter and Dart version to 3.13 and 3.1.
 * Update in_app_purchase_platform_interface to 1.4.0.
 * Add countryCode API.